requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
diff-match-patch==20241021
Django==5.2.8
django-crispy-forms==2.5
django-filter==25.2
django-import-export==4.3.14
djangorestframework==3.16.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.0.0
reportlab==4.4.5
sqlparse==0.5.3
tablib==3.9.0
tzdata==2025.2
psycopg2-binary==2.9.9
asgiref
charset-normalizer
diff-match-patch
Django
django-crispy-forms
django-filter
django-import-export
djangorestframework
et_xmlfile
openpyxl
pillow
psycopg2-binary
reportlab
sqlparse
tablib
tzdata
psycopg2-binary
